requirements.txt
# Core dependencies
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.3.0
apache-airflow-providers-apache-hive==6.1.3
apache-airflow-providers-postgres==5.7.1

# Big Data
pyspark==3.4.1
pyhive==0.7.0
hdfs==2.7.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==1.4.50

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# API Clients
requests==2.31.0
polygon-api-client==1.13.2
alpha-vantage==2.3.1
sec-api==1.0.17

# Data Validation
great-expectations==0.18.3

# Sentiment Analysis
vaderSentiment==3.3.2
textblob==0.17.1
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
retrying==1.3.4
tenacity==8.2.3

# Logging and Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1
